# swagger.yaml
openapi: 3.0.3
info:
  title: Private Services API  - OpenAPI 3.0
  description: "This is the set of api endpoints to support access to Subscripify\
    \ Super Services. Super Services are those that are available to Subscripify's\
    \ super tenants. They provide higher capabilities to super tenants to manage across\
    \ multiple main tenants within the context of the Super-Tenant.<br><br> \nThese\
    \ APIs act as the front end to Private Services engineered and hosted by super\
    \ tenants that need to access the subscripify platform.<br><br>\nExamples of super\
    \ services available only to Super Tenants through this API are: <br>\nTenant\
    \ Management<br>\nSubscription and Plan Management <br>\nBilling and Payments<br><br>\n\
    Subscripify also maintains a set of api endpoints to support access to Subscripify\
    \ Public Services. Subscripify public services are services required by all tenants\
    \ and users on the subscripify platform. All tenants, regardless of type have\
    \ access to these services.<br><br>\nExamples of private services available to\
    \ Super Tenants and to Main Tenants are: <br>\nIdentity Services<br>\nUsage Analytics\
    \ Services<br>\n\nFor more information about Subscripify's tenant architecture\
    \ click here."
  termsOfService: http://www.subscripify.com/terms/apiusage
  contact:
    email: william.ohara@subscripify.com
  license:
    name: Private
    url: http://www.subscripify.com/license
  version: 0.0.1
externalDocs:
  description: Find out more about Subscripify
  url: http://www.subscripify.com
servers:
- url: https://lord-tenant.subscripify.com
tags:
- name: tenants
  description: everything needed to create and manage tenants on the Subscripify platform
  externalDocs:
    description: Link to Tenant Architecture Doc
    url: http://wiki.subscripify.com/tenant-architecture
- name: subscription
  description: everything subscriptions
  externalDocs:
    description: Find out more about our store
    url: http://wiki.subscripify.com/tenant-architecture
paths:
  /tenants/lord:
    post:
      tags:
      - tenants
      summary: Add a new lord tenant
      description: Adds a new lord tenant to the subscripify tenant system. This will
        start a new Kubernetes cluster   on the selected cloud service
      operationId: addLordTenant
      requestBody:
        description: All data required to start a new tenant
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/lordTenantCreateBody'
        required: true
      responses:
        
        "200":
          description: Successful operation returns full tenant data including generated
            felids
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/lordTenantCreateBody'
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "405":
          $ref: "#/components/responses/405"
        "406":
          $ref: "#/components/responses/406"
        "415":
          $ref: "#/components/responses/415"
        "429":
          $ref: "#/components/responses/429"
        default:
          $ref: "#/components/responses/default"
      security:
      - tenant_auth:
        - write:lord-tenants
        - read:lord-tenants
  /tenants/super:
    post:
      tags:
      - tenants
      summary: Add a new super tenant
      description: Adds a new super tenant to the subscripify tenant system. This will
        start a new Kubernetes cluster   on the selected cloud service
      operationId: addSuperTenant
      requestBody:
        description: All data required to start a new tenant
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/superTenantCreateBody'
        required: true
      responses:
        
        "200":
          description: Successful operation returns full tenant data including generated
            felids
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/superTenantCreateBody'
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "405":
          $ref: "#/components/responses/405"
        "406":
          $ref: "#/components/responses/406"
        "415":
          $ref: "#/components/responses/415"
        "429":
          $ref: "#/components/responses/429"
        default:
          $ref: "#/components/responses/default"
      security:
      - tenant_auth:
        - write:lord-tenants
        - read:lord-tenants
        - write:super-tenants
        - read:super-tenants
  /tenants/main:
    post:
      tags:
      - tenants
      summary: Add a new main tenant
      description: Adds a new main tenant to the subscripify tenant system. 
      operationId: addMainTenant
      requestBody:
        description: All data required to start a new tenant
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/mainTenantCreateBody'
        required: true
      responses:  
        "200":
          description: Successful operation returns full tenant data including generated
            felids
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mainTenantCreateBody'
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "405":
          $ref: "#/components/responses/405"
        "406":
          $ref: "#/components/responses/406"
        "415":
          $ref: "#/components/responses/415"
        "429":
          $ref: "#/components/responses/429"
        default:
          $ref: "#/components/responses/default"
      security:
      - tenant_auth:
        - write:lord-tenants
        - read:lord-tenants
        - write:super-tenants
        - read:super-tenants
        - write:main-tenants
        - read:main-tenants
  /tenants/{tenantGuid}:
    get:
      summary: returns a tenant
      description: returns a tenant
      tags:
        - "tenants"
      operationId: getTenant
      parameters:
        - name: tenantGuid
          in: path
          required: true
          schema:
            maxLength: 32
            pattern: "^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b\
              -[0-9a-fA-F]{12}$"
            type: string
            description: The services config UUID to use for the lord tenant. Must
              be a valid services config UUID
      responses:
        '200':
          description: return tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mainTenantCreateBody'
    patch:
      summary: updates a tenant
      description: updates a tenant
      tags:
        - "tenants"
      operationId: updateTenant
      parameters:
        - name: tenantGuid
          in: path
          required: true
          schema:
            maxLength: 32
            pattern: "^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b\
              -[0-9a-fA-F]{12}$"
            type: string
            description: The services config UUID to use for the lord tenant. Must
              be a valid services config UUID
      responses:
        '200':
          description: return tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mainTenantCreateBody'
    delete:
      summary: removes a tenant
      description: removes a tenant
      tags:
        - "tenants"
      operationId: deleteTenant
      parameters:
        - name: tenantGuid
          in: path
          required: true
          schema:
            maxLength: 32
            pattern: "^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b\
              -[0-9a-fA-F]{12}$"
            type: string
            description: The services config UUID to use for the lord tenant. Must
              be a valid services config UUID
      responses:
        '200':
          description: return tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mainTenantCreateBody'
components:
  schemas:
    lordTenantCreateBody:
      required:
      - internalServicesConfig
      - publicServicesConfig
      - superServicesConfig
      - subdomain
      - subscripifyDeploymentCloudLocation
      - tenantType
      type: object
      properties:
        internalServicesConfig:
          maxLength: 32
          pattern: "^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b\
            -[0-9a-fA-F]{12}$"
          type: string
          description: The services config UUID to use for the lord tenant. Must
            be a valid services config UUID
        superServicesConfig:
          maxLength: 32
          pattern: "^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b\
            -[0-9a-fA-F]{12}$"
          type: string
          description: The services config UUID to use for the super-tenant. Must
            be a valid services config UUID
        subdomain:
          maxLength: 30
          pattern: ".[a-z][0-9][-]"
          type: string
          description: The subdomain name string which used for the services namespace
            of the tenant and  providing unique url for each tenant
          example: subscripify
        publicServicesConfig:
          maxLength: 32
          pattern: "^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b\
            -[0-9a-fA-F]{12}$"
          type: string
          description: The services config UUID to use for the main-tenant's public
            services. The services config UUID  used must be a publicServices UUID
            and belong to the liege tenant
        subscripifyDeploymentCloudLocation:
          maxLength: 32
          pattern: azure
          type: string
          description: The cloud provider to deploy to. e.g. The only cloud provider
            supported (currently) is azure.
      additionalProperties: false
    superTenantCreateBody:
      required:
      - superServicesConfig
      - publicServicesConfig
      - subdomain
      - subscripifyDeploymentCloudLocation
      - tenantType
      type: object
      properties:
        superServicesConfig:
          maxLength: 32
          pattern: "^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b\
            -[0-9a-fA-F]{12}$"
          type: string
          description: The services config UUID to use for the super-tenant. Must
            be a valid services config UUID
        tenantType:
          maxLength: 5
          pattern: ^super|lord|main$
          type: string
          description: "Indicate which type of tenant one is spinning up. Main Tenants\
            \ require only a public services config.  Super tenants require a super\
            \ services config and a public config. Lord Tenants require super, public\
            \ and internal services  configs."
        subdomain:
          maxLength: 30
          pattern: ".[a-z][0-9][-]"
          type: string
          description: The subdomain name string which used for the services namespace
            of the tenant and  providing unique url for each tenant
          example: subscripify
        publicServicesConfig:
          maxLength: 32
          pattern: "^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b\
            -[0-9a-fA-F]{12}$"
          type: string
          description: The services config UUID to use for the main-tenant's public
            services. The services config UUID  used must be a publicServices UUID
            and belong to the liege tenant
        subscripifyDeploymentCloudLocation:
          maxLength: 32
          pattern: azure
          type: string
          description: The cloud provider to deploy to. e.g. The only cloud provider
            supported (currently) is azure.
      additionalProperties: false
    mainTenantCreateBody:
      required:
      - publicServicesConfig
      - subdomain
      - subscripifyDeploymentCloudLocation
      - tenantType
      type: object
      properties:
        tenantType:
          maxLength: 5
          pattern: ^super|lord|main$
          type: string
          description: "Indicate which type of tenant one is spinning up. Main Tenants\
            \ require only a public services config.  Super tenants require a super\
            \ services config and a public config. Lord Tenants require super, public\
            \ and internal services  configs."
        subdomain:
          maxLength: 30
          pattern: ".[a-z][0-9][-]"
          type: string
          description: The subdomain name string which used for the services namespace
            of the tenant and  providing unique url for each tenant
          example: subscripify
        publicServicesConfig:
          maxLength: 32
          pattern: "^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b\
            -[0-9a-fA-F]{12}$"
          type: string
          description: The services config UUID to use for the main-tenant's public
            services. The services config UUID  used must be a publicServices UUID
            and belong to the liege tenant
        subscripifyDeploymentCloudLocation:
          maxLength: 32
          pattern: azure
          type: string
          description: The cloud provider to deploy to. e.g. The only cloud provider
            supported (currently) is azure.
      additionalProperties: false
    UnauthorizedError:
      type: object
      properties:
        responseCode:
          type: integer
          example: 401
        message:
          maxLength: 60
          pattern: ".[a-z][0-9]"
          type: string
          example: not authorized to view this resource
      additionalProperties: false
      description: Authentication information is missing or invalid
      example:
        message: not authorized to view this resource
        responseCode: 401
    
  responses:
    "401":
          description: Authorization Required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
    "403":
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
    "405":
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
    "406":
      description: Improper Protocol
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
    "415":
      description: Improper media type
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
    "429":
      description: too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
    default:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
  securitySchemes:
    tenant_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header